refactor: CardNumberInput 컴포넌트 이름을 SeperatedInputList 로 일반화
