+++ src/components/SeperatedInputList/SeperatedInputList.module.scss
@@ -1,4 +1,4 @@
-.card-number-input {
+.seperated-input-list {
   display: inline-flex;
   flex-direction: column;
   width: 320px;
